+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Features
 
 * Introduced mining modes: local, client and server mode. In the client mode (default), miners automatically submit their findings to a central server. Clients don't need to create pull-requests anymore to contribute programs.
+* Change maximum memory usage with the setup command.
 
 ### Enhancements